refactor: optimize navigation structure
